requirements.txt
httpx==0.27.2
boto3==1.35.36
jsonschema==4.23.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
openai==1.51.2
PyYAML==6.0.2
packaging==24.1
pytest==8.3.3
pytest-asyncio==0.24.0
